+++ src/components/lotto/LotteriesDetail.jsx
