+++ src/components/lotto/index.js
