+++ src/services/index.js
@@ -1,2 +1,1 @@
-export { publishLotteries } from './publishLotteries';
 export { getWinningResult } from './getWinningResult';
